Refactor: 코드리뷰 반영-GNB로 이름 변경
